plotted logfc and pval hists

+++ R/6-pseudobulkanddea.R
@@ -50,6 +50,8 @@ main <- function() {
     }
   }
 
+  ### perform DEA and save TMM PB
+  
   for (cohort in cohort_list){
     astrocyte <- paste0("Ast_", cohort, "_SZ.rds")
     excitatory <- paste0("Exc_", cohort, "_SZ.rds")
@@ -57,15 +59,22 @@ main <- function() {
     microglia <- paste0("Mic_", cohort, "_SZ.rds")
     oligodendrocyte <- paste0("Oli_", cohort, "_SZ.rds")
     opc <- paste0("Opc_", cohort, "_SZ.rds")
+    gli <- paste0("Gli_", cohort, "_SZ.rds")
     
-    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc)
+    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc, gli)
     
     for (cell in cell_list) {
       full_path <- paste0("data/data_processed/", cohort, "/Pseudobulk/PseudobulkRaw/", cell)
+      
+      if (!file.exists(full_path)) {
+        message("File ", full_path, " does not exist. Skipping to next.")
+        next
+      }
+      
       pseudobulk <- readRDS(full_path)
       dea_res <- perform_DGE(pseudobulk)
       
-      dea_path <- paste0("results/DEA/", cohort, "/TMMDEAresults_", cell)
+      dea_path <- paste0("results/DEA/", cohort, "/TMM/DEAresults_", cell)
       saveRDS(dea_res$dge_results, dea_path)
       
       tmm_path <- paste0("data/data_processed/", cohort, "/Pseudobulk/PseudobulkTMM/", cell)
@@ -73,6 +82,7 @@ main <- function() {
       }
   }
   
+  ### perform DEA on CPM PB
   for (cohort in cohort_list){
     astrocyte <- paste0("Ast_", cohort, "_SZ.rds")
     excitatory <- paste0("Exc_", cohort, "_SZ.rds")
@@ -80,20 +90,358 @@ main <- function() {
     microglia <- paste0("Mic_", cohort, "_SZ.rds")
     oligodendrocyte <- paste0("Oli_", cohort, "_SZ.rds")
     opc <- paste0("Opc_", cohort, "_SZ.rds")
+    gli <- paste0("Gli_", cohort, "_SZ.rds")
     
-    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc)
+    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc, gli)
     
     for (cell in cell_list) {
       pb_path <- paste0("data/data_processed/", cohort, "/Pseudobulk/PseudobulkCPM/", cell)
+      
+      if (!file.exists(pb_path)) {
+        message("File ", pb_path, " does not exist. Skipping to next.")
+        next
+      }
+      
       pseudobulk <- readRDS(pb_path)
       dea_res <- perform_DGE_on_CPM(pseudobulk)
       
-      dea_path <- paste0("results/DEA/", cohort, "/CPMDEAresults_", cell)
-      saveRDS(dea_res$dge_results, dea_path)
+      dea_path <- paste0("results/DEA/", cohort, "/CPM/DEAresults_", cell)
+      saveRDS(dea_res, dea_path)
     }
   }
   
+  ### CPMLOG PBs and perform DEA
+  for (cohort in cohort_list){
+    astrocyte <- paste0("Ast_", cohort, "_SZ.rds")
+    excitatory <- paste0("Exc_", cohort, "_SZ.rds")
+    inhibitory <- paste0("Inh_", cohort, "_SZ.rds")
+    microglia <- paste0("Mic_", cohort, "_SZ.rds")
+    oligodendrocyte <- paste0("Oli_", cohort, "_SZ.rds")
+    opc <- paste0("Opc_", cohort, "_SZ.rds")
+    gli <- paste0("Gli_", cohort, "_SZ.rds")
+    
+    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc, gli)
+    
+    for (cell in cell_list) {
+      full_path <- paste0("data/data_processed/", cohort, "/Pseudobulk/PseudobulkRaw/", cell)
+      
+      if (!file.exists(full_path)) {
+        message("File ", full_path, " does not exist. Skipping to next.")
+        next
+      }
+      unfiltered <- readRDS(full_path)
+      
+      cpm_matrix <- do_cpm_log(unfiltered$expr, TRUE)
+      final_cpmed <- list(expr = cpm_matrix, meta = unfiltered$meta)
+      
+      cpm_path <- paste0("data/data_processed/", cohort, "/Pseudobulk/PseudobulkCPMLog/", cell)
+      saveRDS(final_cpmed, cpm_path)
+      
+      results <- perform_DGE_on_CPM(final_cpmed)
+      
+      cpmlogres_path <- paste0("results/DEA/", cohort, "/CPMLog/DEAresults_", cell)
+      saveRDS(results, cpmlogres_path)
+    }
+  }
+  
+  ### DEA on CPMLog with other covariates
+  
+  
+  
+  ### create p-value hists
+  
+  for (cohort in cohort_list){
+    astrocyte <- paste0("Ast_", cohort, "_SZ.rds")
+    excitatory <- paste0("Exc_", cohort, "_SZ.rds")
+    inhibitory <- paste0("Inh_", cohort, "_SZ.rds")
+    microglia <- paste0("Mic_", cohort, "_SZ.rds")
+    oligodendrocyte <- paste0("Oli_", cohort, "_SZ.rds")
+    opc <- paste0("Opc_", cohort, "_SZ.rds")
+    gli <- paste0("Gli_", cohort, "_SZ.rds")
+    
+    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc, gli)
+    plot_list <- list()
+    
+    for (cell in cell_list) {
+      res_path <- paste0("results/DEA/", cohort, "/CPM/DEAresults_", cell)
+      
+      cell_type <- strsplit(cell, "_")[[1]][1]
+      
+      if (!file.exists(res_path)) {
+        message("File ", res_path, " does not exist. Skipping to next.")
+        next
+      }
+    results <- readRDS(res_path)
+    
+    res_his <- results |>
+      ggplot(aes(x = PValue)) +
+      geom_histogram() +
+      theme_classic()+
+      labs(
+        title = cell_type,
+        x = "P-value",
+        y = NULL
+      )
+    
+    plot_list[[cell_type]] <- res_his
+    }
+    
+    all_plots <- plot_grid(plotlist = plot_list, ncol = 3, align = "hv")
+    
+    title_plot <- ggdraw() + 
+      draw_label(cohort, x = 0.5, y = 0.5, size = 12, hjust = 0.5)
+    
+    all_plots_with_title <- plot_grid(title_plot, all_plots, ncol = 1, rel_heights = c(0.1, 1))
+    
+    final_path <- paste0("results/DEA/", cohort, "/CPM/pvalueplot_", cohort, ".png")
+    
+    cowplot::save_plot(plot = all_plots_with_title, filename = final_path)
+  }
+  
+  for (cohort in cohort_list){
+    astrocyte <- paste0("Ast_", cohort, "_SZ.rds")
+    excitatory <- paste0("Exc_", cohort, "_SZ.rds")
+    inhibitory <- paste0("Inh_", cohort, "_SZ.rds")
+    microglia <- paste0("Mic_", cohort, "_SZ.rds")
+    oligodendrocyte <- paste0("Oli_", cohort, "_SZ.rds")
+    opc <- paste0("Opc_", cohort, "_SZ.rds")
+    gli <- paste0("Gli_", cohort, "_SZ.rds")
+    
+    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc, gli)
+    plot_list <- list()
+    
+    for (cell in cell_list) {
+      res_path <- paste0("results/DEA/", cohort, "/CPMLog/DEAresults_", cell)
+      
+      cell_type <- strsplit(cell, "_")[[1]][1]
+      
+      if (!file.exists(res_path)) {
+        message("File ", res_path, " does not exist. Skipping to next.")
+        next
+      }
+      results <- readRDS(res_path)
+      
+      res_his <- results |>
+        ggplot(aes(x = PValue)) +
+        geom_histogram() +
+        theme_classic()+
+        labs(
+          title = cell_type,
+          x = "P-value",
+          y = NULL
+        )
+      
+      plot_list[[cell_type]] <- res_his
+    }
+    
+    all_plots <- plot_grid(plotlist = plot_list, ncol = 3, align = "hv")
+    
+    title_plot <- ggdraw() + 
+      draw_label(cohort, x = 0.5, y = 0.5, size = 12, hjust = 0.5)
+    
+    all_plots_with_title <- plot_grid(title_plot, all_plots, ncol = 1, rel_heights = c(0.1, 1))
+    
+    final_path <- paste0("results/DEA/", cohort, "/CPMLog/pvalueplot_", cohort, ".png")
+    
+    cowplot::save_plot(plot = all_plots_with_title, filename = final_path)
+  }
+  
+  for (cohort in cohort_list){
+    astrocyte <- paste0("Ast_", cohort, "_SZ.rds")
+    excitatory <- paste0("Exc_", cohort, "_SZ.rds")
+    inhibitory <- paste0("Inh_", cohort, "_SZ.rds")
+    microglia <- paste0("Mic_", cohort, "_SZ.rds")
+    oligodendrocyte <- paste0("Oli_", cohort, "_SZ.rds")
+    opc <- paste0("Opc_", cohort, "_SZ.rds")
+    gli <- paste0("Gli_", cohort, "_SZ.rds")
+    
+    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc, gli)
+    plot_list <- list()
+    
+    for (cell in cell_list) {
+      res_path <- paste0("results/DEA/", cohort, "/TMM/DEAresults_", cell)
+      
+      cell_type <- strsplit(cell, "_")[[1]][1]
+      
+      if (!file.exists(res_path)) {
+        message("File ", res_path, " does not exist. Skipping to next.")
+        next
+      }
+      results <- readRDS(res_path)
+      
+      res_his <- results |>
+        ggplot(aes(x = PValue)) +
+        geom_histogram() +
+        theme_classic()+
+        labs(
+          title = cell_type,
+          x = "P-value",
+          y = NULL
+        )
+      
+      plot_list[[cell_type]] <- res_his
+    }
+    
+    all_plots <- plot_grid(plotlist = plot_list, ncol = 3, align = "hv")
+    
+    title_plot <- ggdraw() + 
+      draw_label(cohort, x = 0.5, y = 0.5, size = 12, hjust = 0.5)
+    
+    all_plots_with_title <- plot_grid(title_plot, all_plots, ncol = 1, rel_heights = c(0.1, 1))
+    
+    final_path <- paste0("results/DEA/", cohort, "/TMM/pvalueplot_", cohort, ".png")
+    
+    cowplot::save_plot(plot = all_plots_with_title, filename = final_path)
+  }
+  
+  ###create hists for logFC
+  for (cohort in cohort_list){
+    astrocyte <- paste0("Ast_", cohort, "_SZ.rds")
+    excitatory <- paste0("Exc_", cohort, "_SZ.rds")
+    inhibitory <- paste0("Inh_", cohort, "_SZ.rds")
+    microglia <- paste0("Mic_", cohort, "_SZ.rds")
+    oligodendrocyte <- paste0("Oli_", cohort, "_SZ.rds")
+    opc <- paste0("Opc_", cohort, "_SZ.rds")
+    gli <- paste0("Gli_", cohort, "_SZ.rds")
+    
+    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc, gli)
+    plot_list <- list()
+    
+    for (cell in cell_list) {
+      res_path <- paste0("results/DEA/", cohort, "/CPM/DEAresults_", cell)
+      
+      cell_type <- strsplit(cell, "_")[[1]][1]
+      
+      if (!file.exists(res_path)) {
+        message("File ", res_path, " does not exist. Skipping to next.")
+        next
+      }
+      results <- readRDS(res_path)
+      
+      res_his <- results |>
+        ggplot(aes(x = logFC)) +
+        geom_histogram() +
+        theme_classic()+
+        labs(
+          title = cell_type,
+          x = "logFC",
+          y = NULL
+        )
+      
+      plot_list[[cell_type]] <- res_his
+    }
+    
+    all_plots <- plot_grid(plotlist = plot_list, ncol = 3, align = "hv")
+    
+    title_plot <- ggdraw() + 
+      draw_label(cohort, x = 0.5, y = 0.5, size = 12, hjust = 0.5)
+    
+    all_plots_with_title <- plot_grid(title_plot, all_plots, ncol = 1, rel_heights = c(0.1, 1))
+    
+    final_path <- paste0("results/DEA/", cohort, "/CPM/logFCplot_", cohort, ".png")
+    
+    cowplot::save_plot(plot = all_plots_with_title, filename = final_path)
+  }
+  
+  for (cohort in cohort_list){
+    astrocyte <- paste0("Ast_", cohort, "_SZ.rds")
+    excitatory <- paste0("Exc_", cohort, "_SZ.rds")
+    inhibitory <- paste0("Inh_", cohort, "_SZ.rds")
+    microglia <- paste0("Mic_", cohort, "_SZ.rds")
+    oligodendrocyte <- paste0("Oli_", cohort, "_SZ.rds")
+    opc <- paste0("Opc_", cohort, "_SZ.rds")
+    gli <- paste0("Gli_", cohort, "_SZ.rds")
+    
+    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc, gli)
+    plot_list <- list()
+    
+    for (cell in cell_list) {
+      res_path <- paste0("results/DEA/", cohort, "/CPMLog/DEAresults_", cell)
+      
+      cell_type <- strsplit(cell, "_")[[1]][1]
+      
+      if (!file.exists(res_path)) {
+        message("File ", res_path, " does not exist. Skipping to next.")
+        next
+      }
+      results <- readRDS(res_path)
+      
+      res_his <- results |>
+        ggplot(aes(x = logFC)) +
+        geom_histogram() +
+        theme_classic()+
+        labs(
+          title = cell_type,
+          x = "logFC",
+          y = NULL
+        )
+      
+      plot_list[[cell_type]] <- res_his
+    }
+    
+    all_plots <- plot_grid(plotlist = plot_list, ncol = 3, align = "hv")
+    
+    title_plot <- ggdraw() + 
+      draw_label(cohort, x = 0.5, y = 0.5, size = 12, hjust = 0.5)
+    
+    all_plots_with_title <- plot_grid(title_plot, all_plots, ncol = 1, rel_heights = c(0.1, 1))
+    
+    final_path <- paste0("results/DEA/", cohort, "/CPMLog/logFCplot_", cohort, ".png")
+    
+    cowplot::save_plot(plot = all_plots_with_title, filename = final_path)
+  }
+  
+  for (cohort in cohort_list){
+    astrocyte <- paste0("Ast_", cohort, "_SZ.rds")
+    excitatory <- paste0("Exc_", cohort, "_SZ.rds")
+    inhibitory <- paste0("Inh_", cohort, "_SZ.rds")
+    microglia <- paste0("Mic_", cohort, "_SZ.rds")
+    oligodendrocyte <- paste0("Oli_", cohort, "_SZ.rds")
+    opc <- paste0("Opc_", cohort, "_SZ.rds")
+    gli <- paste0("Gli_", cohort, "_SZ.rds")
+    
+    cell_list <- c(astrocyte, excitatory, inhibitory, microglia, oligodendrocyte, opc, gli)
+    plot_list <- list()
+    
+    for (cell in cell_list) {
+      res_path <- paste0("results/DEA/", cohort, "/TMM/DEAresults_", cell)
+      
+      cell_type <- strsplit(cell, "_")[[1]][1]
+      
+      if (!file.exists(res_path)) {
+        message("File ", res_path, " does not exist. Skipping to next.")
+        next
+      }
+      results <- readRDS(res_path)
+      
+      res_his <- results |>
+        ggplot(aes(x = logFC)) +
+        geom_histogram() +
+        theme_classic()+
+        labs(
+          title = cell_type,
+          x = "logFC",
+          y = NULL
+        )
+      
+      plot_list[[cell_type]] <- res_his
+    }
+    
+    all_plots <- plot_grid(plotlist = plot_list, ncol = 3, align = "hv")
+    
+    title_plot <- ggdraw() + 
+      draw_label(cohort, x = 0.5, y = 0.5, size = 12, hjust = 0.5)
+    
+    all_plots_with_title <- plot_grid(title_plot, all_plots, ncol = 1, rel_heights = c(0.1, 1))
+    
+    final_path <- paste0("results/DEA/", cohort, "/TMM/logFCplot_", cohort, ".png")
+    
+    cowplot::save_plot(plot = all_plots_with_title, filename = final_path)
+  }
+  
 }
+
+### helper functions
 
 create_pseudo_bulk <- function(expr, meta) {
  
@@ -180,8 +528,29 @@ perform_DGE_on_CPM <- function(PB) {
   
   design <- model.matrix(~ group, data = PB$meta)
   y <- DGEList(counts = PB$expr, group = PB$meta$group)
-  y$samples$norm.factors <- 1
  
+  x <- estimateDisp(y, design, trend.method = "locfit", tagwise = TRUE, prior.df = NULL)   # Estimate dispersion
+  fit <- glmFit(x, design = design)   # Fit GLM (edgeR)
+  test <- glmLRT(fit)   # Likelihood ratio test
+  
+  # Extract results
+  res <- topTags(test, n = Inf) %>%
+    as.data.frame() %>%
+    rownames_to_column('gene') %>%
+    mutate(test = 'pseudobulk_edgeR')
+  
+  return(res)
+}
+
+perform_DGE_with_covs <- function(PB) {
+  # Check if all sample names in the expression matrix match group sample names
+  if (!all(colnames(PB$expr) %in% PB$meta$group_sample)) {
+    stop("Sample names in expression matrix do not match group sample names in meta data")
+  }
+  
+  design <- model.matrix(~ group + sex + age, data = PB$meta)
+  y <- DGEList(counts = PB$expr, group = PB$meta$group)
+  
   x <- estimateDisp(y, design, trend.method = "locfit", tagwise = TRUE, prior.df = NULL)   # Estimate dispersion
   fit <- glmFit(x, design = design)   # Fit GLM (edgeR)
   test <- glmLRT(fit)   # Likelihood ratio test
